chore(signup): add jwt to user after signup [Delivers #166871320]
