+++ api/routes/users.js
@@ -10,7 +10,4 @@ router.get('/', userController.fetchUser);
 
 router.post('/signup', userController.createUser);
 
-router.post('/login', userController.loginUser);
-
-
 export default router;
