+++ models/db/userDb.js
@@ -1,10 +1,11 @@
 const users = [{
   id: 1,
   email: 'me@gmail.com',
-  firstName: 'Jacob',
-  lastName: 'Lawson',
+  first_name: 'Jacob',
+  last_name: 'Lawson',
   password: 'Jabowe1',
   address: 'No 1 Adebowale crescent lekki, Lagos',
+  phone_number: '070-622-78182',
   is_admin: true,
 
 }];
